# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Slimlytics API",
    "version": "1.0.0",
    "description": "Account, personal API token, site configuration, privacy-preserving collection, reporting, goals, export, and real-time APIs for Slimlytics.\n\nUse a short-lived session JWT for creating personal API tokens. Use a `slyt_\u2026` personal API token for durable CLI and agent access. Both use `Authorization: Bearer <token>`. Collection write keys authorize ingestion only.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://slimlytics.com",
      "description": "Slimlytics Cloud"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/billing": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Get the account's plan and usage",
        "operationId": "getBilling",
        "description": "Session login only; API and agent tokens are refused (403).",
        "responses": {
          "200": {
            "description": "Plan, subscription, and usage",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "enabled"
                  ],
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    },
                    "plan": {
                      "type": "object",
                      "required": [
                        "id",
                        "name"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "sites": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "description": "Maximum owned sites; null is unlimited"
                        },
                        "dailyPageViews": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "description": "Soft daily limit across the account's sites; null is unlimited"
                        },
                        "monthlyPriceCents": {
                          "type": "integer"
                        },
                        "annualPriceCents": {
                          "type": "integer"
                        },
                        "currency": {
                          "type": "string"
                        }
                      }
                    },
                    "planSource": {
                      "type": "string",
                      "enum": [
                        "default",
                        "stripe",
                        "admin"
                      ]
                    },
                    "subscriptionStatus": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "interval": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "month",
                        "year",
                        null
                      ]
                    },
                    "currentPeriodEnd": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "usage": {
                      "type": "object",
                      "properties": {
                        "sites": {
                          "type": "integer"
                        },
                        "pageViewsToday": {
                          "type": "integer"
                        }
                      }
                    },
                    "plans": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "sites": {
                            "type": [
                              "integer",
                              "null"
                            ],
                            "description": "Maximum owned sites; null is unlimited"
                          },
                          "dailyPageViews": {
                            "type": [
                              "integer",
                              "null"
                            ],
                            "description": "Soft daily limit across the account's sites; null is unlimited"
                          },
                          "monthlyPriceCents": {
                            "type": "integer"
                          },
                          "annualPriceCents": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "checkoutAvailable": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/billing/checkout": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Start a Stripe Checkout subscription",
        "operationId": "createBillingCheckout",
        "description": "Session login only. Existing subscribers receive a Customer Portal URL instead, so plans change without a second subscription.",
        "responses": {
          "200": {
            "description": "Checkout or portal URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "url"
                  ],
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "portal": {
                      "type": "boolean",
                      "description": "True when an existing subscriber is sent to the Customer Portal instead"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "plan"
                ],
                "properties": {
                  "plan": {
                    "type": "string"
                  },
                  "interval": {
                    "type": "string",
                    "enum": [
                      "month",
                      "year"
                    ],
                    "default": "month"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/billing/portal": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Open the Stripe Customer Portal",
        "operationId": "createBillingPortal",
        "description": "Session login only. Manage payment methods, change plans, or cancel.",
        "responses": {
          "200": {
            "description": "Portal URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "url"
                  ],
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "portal": {
                      "type": "boolean",
                      "description": "True when an existing subscriber is sent to the Customer Portal instead"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/billing/webhook": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Stripe webhook",
        "operationId": "stripeWebhook",
        "description": "Called by Stripe. Requires a valid Stripe-Signature header; each event is applied once and re-syncs the customer's subscription.",
        "responses": {
          "200": {
            "description": "Event received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": {
                      "type": "boolean"
                    },
                    "duplicate": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid signature or payload"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or slyt_ personal API token",
        "description": "Session JWT or personal API token. Token creation specifically requires a session JWT."
      },
      "sessionQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Short-lived session JWT for native EventSource clients. Never put personal API tokens in URLs."
      },
      "serverWriteKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Slimlytics-Server-Key",
        "description": "Rotatable site secret accepted only by server request ingestion."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "Forbidden": {
        "description": "Insufficient site role",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "bad_request",
                  "rate_limited",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}